+++ Configuration/RequestMiddlewares.php
@@ -2,6 +2,7 @@
 
 declare(strict_types=1);
 
+use AOE\Crawler\Middleware\ContentFinisher;
 use AOE\Crawler\Middleware\CrawlerInitialization;
 use AOE\Crawler\Middleware\FrontendUserAuthenticator;
 
@@ -21,6 +22,12 @@ return [
             'before' => [
                 'typo3/cms-core/normalizedParams',
             ],
+        ],
+        'aoe/crawler/content-finisher' => [
+            'target' => ContentFinisher::class,
+            'before' => [
+                'typo3/cms-frontend/timetracker'
+            ]
         ],
     ],
 ];